Unify CLI entrypoints to fix project keryx.ts command parity (#200)
Extract a shared buildProgram() function in packages/keryx/util/cli.ts that registers all CLI commands (new, upgrade, generate, start, actions). Both the framework's keryx.ts and scaffolded projects now call this function to ensure consistent CLI experience. Projects can now run `bun keryx.ts upgrade` and all other framework commands—previously only `start` and `actions` worked locally.

Also remove commander from scaffolded package.json dependencies since projects no longer import it directly, and bump framework version from 0.9.0 to 0.10.0.
