+++ packages/keryx/package.json
@@ -1,6 +1,6 @@
 {
   "name": "keryx",
-  "version": "0.9.0",
+  "version": "0.10.0",
   "module": "index.ts",
   "type": "module",
   "license": "MIT",
