+++ packages/keryx/templates/scaffold/keryx.ts.mustache
@@ -1,55 +1,15 @@
 #! /usr/bin/env bun
 
 // Set rootDir before any framework code loads actions
-import { api } from "./index";
+import "./index";
 
-import { Command } from "commander";
-import { Action, globLoader } from "keryx";
-import { addActionToProgram } from "keryx/util/cli.ts";
-import path from "path";
+import { buildProgram } from "keryx/util/cli.ts";
 import pkg from "./package.json";
 
-const program = new Command();
-program.name(pkg.name).description(pkg.description).version(pkg.version);
-
-program
-  .command("start")
-  .summary("Run the server")
-  .description("Start the Keryx server")
-  .action(async () => {
-    await api.start();
-  });
-
-// Load actions from the project directory
-let actions: Action[] = [];
-try {
-  actions = await globLoader<Action>(path.join(api.rootDir, "actions"));
-} catch {
-  // project may not have actions yet
-}
-actions.forEach((action) => addActionToProgram(program, action));
-
-program
-  .command("actions")
-  .summary("List all actions")
-  .action(async () => {
-    const actionSpacing =
-      actions.map((a) => a.name.length).reduce((a, b) => Math.max(a, b), 0) +
-      2;
-    const routeSpacing =
-      actions
-        .map((a) =>
-          a.web ? a.web.route.toString().length + a.web.method.length : 0,
-        )
-        .reduce((a, b) => Math.max(a, b), 0) + 2;
-
-    actions
-      .sort((a, b) => a.name.localeCompare(b.name))
-      .forEach((action) => {
-        console.log(
-          `${action.name}${" ".repeat(actionSpacing - action.name.length)} ${action.web ? `[${action.web.method}] ${action.web.route}` : " "}${" ".repeat(routeSpacing - (action.web ? action.web.method.length + action.web.route.toString().length + 2 : 0))} ${action.description ?? ""}`,
-        );
-      });
-  });
+const program = await buildProgram({
+  name: pkg.name,
+  description: pkg.description,
+  version: pkg.version,
+});
 
 program.parse();
